+++ packages/terra-site/src/TestLinks.jsx
@@ -37,6 +37,7 @@ const TestLinks = () => (
       <li><Link to="/tests/modal-tests">Modal Tests</Link></li>
       <li><Link to="/tests/modal-manager-tests">Modal Manager Tests</Link></li>
       <li><Link to="/tests/multi-select-list-tests">MultiSelectList Tests</Link></li>
+      <li><Link to="/tests/multi-select-table-tests">MultiSelectTable Tests</Link></li>
       <li><Link to="/tests/overlay-tests">Overlay Tests</Link></li>
       <li><Link to="/tests/overlay-container-tests">OverlayContainer Tests</Link></li>
       <li><Link to="/tests/popup-tests">Popup Tests</Link></li>